requirements.txt
streamlit==1.42.0
supabase==2.12.0
yfinance==0.2.54
pandas==2.2.3
langchain-groq==0.2.4
langgraph==0.2.73
python-dotenv==1.0.1
pydantic==2.10.6